#1 make joda-time dependency optional and only add converter if present on class path
